Revert ""updatebatch lambda""
This reverts commit 784feed59445fcc8017363e3d5f94073768604b3.
