+++ src/batch/getBatchesByCurriculumFn/getBatchesByCurriculum.ts
@@ -12,14 +12,7 @@ export default async function handler(event: APIGatewayProxyEvent) {
     }
     const curriculumId = (event.pathParameters.curriculumId)
 
-    // Attempt to establish a connection; in the case of a failure, give
-    //  error code 500
-    try {
-        await client.connect();
-    } catch (err) {
-        console.log(err)
-        return new HTTPResponse(500, "Unable to Connect to the Database")
-    }
+
 
     const data = [curriculumId];
     let res;
@@ -29,12 +22,10 @@ export default async function handler(event: APIGatewayProxyEvent) {
         res = await client.query(text, data)
     } catch (err) {
         console.log(err);
-        await client.end()
         return new HTTPResponse(400, "Unable to Query the information")
     }
 
     // If all went well, returns everything that the query retrieved
-    await client.end()
     console.log(res.rows);
     return new HTTPResponse(200, res.rows)
 };
