+++ src/batch/removeBatchFn/removeBatch.test.ts
@@ -1,31 +1,31 @@
 jest.mock('../../global/postgres')
 import client from '../../global/postgres'
-import { batch } from '../../global/mockTable'
+import { trainer } from '../../global/mockTable'
 import { APIGatewayProxyEvent } from 'aws-lambda';
-import handler from './getBatchById';
+import handler from './removeBatch';
+const input: unknown = { pathParameters: { batchId : 23 } }
 
-const input: unknown = { pathParameters: batch }
-describe('Testing getBatchById Handler', () => {
+describe('Testing DeleteTrainer  Handler', () => {
     it('should succeed with status code 200', async () => {
         (client.query as jest.Mock).mockImplementationOnce(() => {
-            return { rows: batch }
+            return { rows: trainer } // look into mockReturn
         })
         const res = await handler(input as APIGatewayProxyEvent);
         expect(res.statusCode).toEqual(200);
     })
 
-    it('should fail with 400, from a pathparamaters is null', async () => {
+    it('should fail with 400, from a pathparmaters is null', async () => {
         const res = await handler({} as APIGatewayProxyEvent)
         expect(res.statusCode).toEqual(400);
     })
 
-    it('should fail with 500, from a database connection error', async () => {
+/*     it('should fail with 500, from a database connection error', async () => {
         (client.connect as jest.Mock).mockImplementationOnce(() => {
             throw "error"
         })
         const res = await handler(input as APIGatewayProxyEvent)
         expect(res.statusCode).toEqual(500);
-    })
+    }) */
 
     it('should fail with 400, from a database query error', async () => {
         (client.query as jest.Mock).mockImplementationOnce(() => {
