+++ src/batch/removeBatchFn/removeBatch.ts
@@ -11,13 +11,7 @@ export default async function handler(event: APIGatewayProxyEvent) {
     }
     const batchId = event.pathParameters.batchId;
     //Try connecting to the Database
-    try {
-        await client.connect();
 
-    } catch (err) {
-        console.log(err)
-        return new HTTPResponse(500, "Unable to Connect to the DataBase")
-    }
     let res;
     //Try querying the DataBase
     try {
@@ -25,10 +19,8 @@ export default async function handler(event: APIGatewayProxyEvent) {
 
     } catch (err) {
         console.log(err);
-        await client.end()
         return new HTTPResponse(400, "Unable to Query the information")
     }
     //Return the created trainer
-    await client.end()
     return new HTTPResponse(200, res.rows)
 };
